Move invalid.token.interceptor to api module
